+++ src/vgabuff.rs
@@ -1,3 +1,5 @@
+use crate::{print, println};
+
 use core::fmt;
 use lazy_static::lazy_static;
 use spin::Mutex;